feat(rxfetch(es)): First Commit
Added my versions of rxfetch (no_brew is primarily used since the brew list command takes a whiiiile). Added a couple functions to determine whether JankyBorders is installed, whether it's configured via a bordersrc file, and whether its running (as a brew service). If all  are true then the border of the fetch is set to the active_color. This works for both gradient types, glow, and normal.