feat(rxfetch(es)): Border + term
Forgot to change the other two printed_term instances for the term function; meant that when using iterm2 the term line was blank. Also forgot to add the call to the jankyborders function in the main rxfetch script so the border didn't have any color defined. Both should be good now.